requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-allauth==0.49.0
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-tinymce==3.4.0
et-xmlfile==1.1.0
idna==3.3
lxml==4.8.0
markdown2==2.4.2
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
pyquery==1.4.3
python-decouple==3.6
python-dotenv==0.19.2
python3-openid==3.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
